+++ src/App.tsx
@@ -3,6 +3,7 @@ import './App.css'
 import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
 import getGameData from './data/index'
+import Game from './pages/Game/Game'
 import Home from './pages/Home/Home'
 
 function App() {
@@ -12,8 +13,8 @@ function App() {
     <BrowserRouter basename={`/${process.env.PUBLIC_URL}`}>
       <Routes>
         <Route path="/" element={<Home data={gameData} />} />
-        {/* <Route path="/:gameKey" element={<Game data={gameData} />} />
-        <Route path="/:gameKey/:runKey" element={<Run data={gameData} />} />
+        <Route path="/:gameKey" element={<Game data={gameData} />} />
+        {/* <Route path="/:gameKey/:runKey" element={<Run data={gameData} />} />
         <Route path="/:gameKey/:runKey/:teamKey" element={<Team data={gameData} />} />
         <Route path="locales/en/translation.json"></Route> */}
       </Routes>
