+++ src/components/PageComponent/PageTitle.tsx
@@ -1,6 +1,6 @@
-import { KeyObject, RunGameData, TeamGameData, TravelBingoGameData } from '../../data/interfaces'
-import t2 from '../../utils/t2'
-import { DynamicIconComponent } from '../DynamicIcon/DynamicIcon'
+import DynamicIcon from 'components/DynamicIcon/DynamicIcon'
+import { KeyObject, RunGameData, TeamGameData, TravelBingoGameData } from 'data/interfaces'
+import t2 from 'utils/t2'
 
 interface Props {
   game: TravelBingoGameData & KeyObject
@@ -11,7 +11,7 @@ interface Props {
 function PageTitle({ game, run, team }: Props) {
   return (
     <div style={{ display: 'flex', alignItems: 'center' }}>
-      <DynamicIconComponent iconName={game.icon} style={{ marginRight: 8, fontSize: '32px' }} />
+      <DynamicIcon iconName={game.icon} style={{ marginRight: 8, fontSize: '32px' }} />
       <h2 style={{ margin: 0 }}>{team ? t2(team.name) : run ? t2(run.name) : t2(game.title)}</h2>
     </div>
   )
